requirements.txt
amqp==2.4.1
billiard==3.6.0.0
celery==4.2.0
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.4
django-admin-tools==0.8.1
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-debug-toolbar==1.11
django-filter==2.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-import-export==1.2.0
django-js-asset==1.2.1
django-mptt==0.9.1
django-redis==4.10.0
django-taggit==0.23.0
django-watson==1.5.2
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
kombu==4.3.0
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
PyYAML==3.13
redis==3.1.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
vine==1.2.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
